use same script for both commandline and web app.
